+++ src/m_cheat.c
@@ -72,8 +72,8 @@ dboolean cht_CheckCheat(cheatseq_t *cht, unsigned char key)
         return true;
     }
 
-    // [BH] you have 4 seconds to enter all characters of a cheat sequence correctly
-    if (cht->timeout && leveltime - cht->timeout > HU_MSGTIMEOUT)
+    // [BH] you have 2 or 4 seconds to enter all characters of a cheat sequence correctly
+    if (cht->timeout && leveltime - cht->timeout > (cht->longtimeout ? HU_MSGTIMEOUT : HU_MSGTIMEOUT / 2))
     {
         cht->chars_read = 0;
         cht->param_chars_read = 0;